% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotel_reviews.R
\docType{data}
\name{hotel_reviews}
\alias{hotel_reviews}
\title{Hotel Reviews}
\format{
A data frame with 38932 rows and 5 variables:
\describe{
\item{\code{User_ID}}{character unique ID of the customer}
\item{\code{Description}}{character description of the review posted}
\item{\code{Browser_Used}}{character browser used to post the review}
\item{\code{Device_Used}}{character device used to post the review}
\item{\code{Is_Response}}{character target Variable}
}
}
\source{
\url{https://www.kaggle.com/datasets/harmanpreet93/hotelreviews}
}
\usage{
hotel_reviews
}
\description{
Dataset from a leading travel site that contains hotel reviews by customers
}
\keyword{datasets}
