% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMiss.R
\name{imputeMiss}
\alias{imputeMiss}
\title{Impute Missing Data in  a Dataset}
\usage{
imputeMiss(df, method = "mean", method_cat = "frequent")
}
\arguments{
\item{df}{A dataframe}

\item{method}{Method to replace missing data in numerical observations, default is mean, but also can choose "median"}

\item{method_cat}{Method to replace missing data in categorical observations, default is most frequent, can also select "sample" for random sample.}
}
\value{
A datafame with no missing values in any observation
}
\description{
Impute a dataframe, with different values depending on proportion and class of observations missing.
If a categorical variable is not a factor, character, or logical observation, the function
will not impute.
}
\examples{
\dontrun{
## from 'qacbase' package
# library(qacBase)
# data(tv)
# imputeMiss(tv)
# imputeMiss(tv, method = "median", method_cat = "sample")}
}
